==> prototype_ver_01/2) Extract actor adjective.ipynb <==
## 1. 패키지 로드

import numpy as np
import pandas as pd
from konlpy.tag import Okt
from collections import Counter
from tqdm import tqdm_notebook

## 2. 데이터 로드

actor_df = pd.read_csv('../text_data/Actor.csv', encoding='utf-8-sig')
actor_df.head(2)

## 3. 불용어 선정을 위한 데이터 전처리 및 분석

### 1) 데이터 전처리

# 한글 외에 모두 제거
actor_df['activation_content'] = actor_df['activation_content'].str.replace('[^가-힣 ]', '')
actor_df['배역내용'] = actor_df['배역내용'].str.replace('[^가-힣 ]', '')
actor_df['제목, 내용'] = actor_df['제목, 내용'].str.replace('[^가-힣 ]', '')

# Okt 객체 선언
okt = Okt()

# 토큰화된 morphs를 각각 담을 list
activation_tag = []
act_tag = []
community_tag = []

# 형태소 토큰화
for activation in tqdm_notebook(actor_df['activation_content']):
    morph = okt.pos(activation)
    activation_tag.append(morph)

for act in tqdm_notebook(actor_df['배역내용']):
    morph = okt.pos(act)
    act_tag.append(morph)
    
for community in tqdm_notebook(actor_df['제목, 내용']):
    morph = okt.pos(community)
    community_tag.append(morph)

# 형용사만 추출
stopwords = []
activation_bucket_list = []
act_bucket_list = []
community_bucket_list = []

for activation in activation_tag:
    for word, tag in activation:
        if tag in ['Adjective']:
            if word not in stopwords:
                activation_bucket_list.append(word)

for act in act_tag:
    for word, tag in act:
        if tag in ['Adjective']:
            if word not in stopwords:
                act_bucket_list.append(word)
                
for community in community_tag:
    for word, tag in community:
        if tag in ['Adjective']:
            if word not in stopwords:
                community_bucket_list.append(word)

### 2) 불용어 선정

# 불용어 선정을 위한 counter
activation_counts = Counter(activation_bucket_list)
activation_counts.most_common(50)

act_counts = Counter(act_bucket_list)
act_counts.most_common(50)

community_counts = Counter(community_bucket_list)
community_counts.most_common(50)

# 상위 50개를 모두 불용어로 지정
for stopword, _ in activation_counts.most_common(50):
    stopwords.append(stopword)

for stopword, _ in act_counts.most_common(50):
    stopwords.append(stopword)

for stopword, _ in community_counts.most_common(50):
    stopwords.append(stopword)

# 선정된 불용어 확인
stopwords = set(stopwords) # 중복 제거
stopwords

## 4. 형용사 추출 (불용어 제외)

# 형용사만 추출
activation_total_keywords = []
act_total_keywords = []
community_total_keywords = []
activation_bucket_list = []
act_bucket_list = []
community_bucket_list = []

for activation in activation_tag:
    for word, tag in activation:
        if tag in ['Adjective']:
            if word not in stopwords:
                activation_bucket_list.append(word)
    activation_bucket_list = list(set(activation_bucket_list))
    activation_total_keywords.append(", ".join(activation_bucket_list))

for act in act_tag:
    for word, tag in act:
        if tag in ['Adjective']:
            if word not in stopwords:
                act_bucket_list.append(word)
    act_bucket_list = list(set(act_bucket_list))
    act_total_keywords.append(", ".join(act_bucket_list))
                
for community in community_tag:
    for word, tag in community:
        if tag in ['Adjective']:
            if word not in stopwords:
                community_bucket_list.append(word)
    community_bucket_list = list(set(community_bucket_list))
    community_total_keywords.append(", ".join(community_bucket_list))

# 키워드 합치기
total_keywords = []

for i in range(len(activation_total_keywords)):
    add_list = activation_total_keywords[i] + act_total_keywords[i] + community_total_keywords[i]
    total_keywords.append(add_list)

# 키워드 중복제거
duplicate_total_keywords = []

for i in range(len(total_keywords)):
    tmp = total_keywords[i].split(', ')
    tmp = list(set(tmp))
    duplicate_total_keywords.append(', '.join(tmp))

## 5. 데이터 프레임에 합치기

actor_df['keywords'] = duplicate_total_keywords
actor_df.head(2)

## 6. 데이터 프레임 저장하기

actor_df.to_csv('../text_data/actor_data_keyword.csv', encoding='utf-8-sig', index=False)
