==> prototype_ver_01/3) Extract webtoon adjective.ipynb <==
## 1. 패키지 로드

import pandas as pd
from konlpy.tag import Okt
from collections import Counter
from tqdm import tqdm_notebook

## 2. 데이터 로드

webtoon_df = pd.read_csv('../text_data/Webtoon.csv')
webtoon_df.head()

## 3. 불용어 지정을 위한 데이터 분석

# Okt 객체 선언
okt = Okt()

# 형태소 토큰화
sentences_tag = []

for sentence in tqdm_notebook(webtoon_df['설명']):
    morph = okt.pos(sentence)
    sentences_tag.append(morph)

# 형용사만 추출
stopwords = []
bucket_list = []

for my_sentence in sentences_tag:
    for word, tag in my_sentence:
        if tag in ['Adjective']:
            if word not in stopwords:
                bucket_list.append(word)

# 불용어 선정을 위한 counter
counts = Counter(bucket_list)
counts.most_common(50)

# 불용어 지정
stopwords = ['먼', '그러다', '다를', '있자', '있는줄', '그럼에도', '뻔하다가', '새로','미를', '미에게', '있다', '있는', '같은', '없는', '그런', '상당히', '아닌', '좋은', '인해', '없다', '이런', '있었다', '아니라', '많은', '있으며', '의해', '있다는','가능한', '이런건', '같긴' '있지만', '있던', '있다고', '같다', '아니다', '있을', '있고', '없다고', '굉장히', '어떻게', '많다', '있었던', '의하면','당했', '아무런', '아니였냐', '완전히', '있는데', '다르게', '어떤', '있어', '당한다', '당해', '안된다', '있어서', '당하는', '상당한', '말라', '그래', '당하고', '아니고', '없고', '당했는데', '없기', '없지만', '꼬이기', '그렇지']

## 4. 형용사 추출

total_keyword = [] 
sentences_tag = []

for sentence in tqdm_notebook(webtoon_df['설명']):
    morph = okt.pos(sentence)
    sentences_tag.append(morph)
        
for my_sentence in tqdm_notebook(sentences_tag):
    bucket_list = []
    for word, tag in my_sentence:
        if tag in ['Adjective']:
            if word not in stopwords:
                bucket_list.append(word)
    total_keyword.append(", ".join(bucket_list))

# 등장인물 별로 추출된 형용사가 담긴 list
total_keyword[:5]

## 5. 형용사 data 병합

webtoon_df['keyword'] = total_keyword
webtoon_df.head(3)

## 6. 형용사가 추가된 데이터 프레임 저장

webtoon_df.to_csv('../text_data/webtoon_data_keyword.csv', index=False, encoding='utf-8-sig')
