==> prototype_ver_01/4) Extract keyword from actor and webtoon.ipynb <==
## 1. 패키지 로드

import pandas as pd
import numpy as np
import re
from tqdm import tqdm_notebook

from konlpy.tag import Okt
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from sentence_transformers import SentenceTransformer
import itertools

## 2. 데이터 로드

actor_df = pd.read_csv('../text_data/actor_data_keyword.csv', encoding='utf-8-sig')
webtoon_df = pd.read_csv('../text_data/webtoon_data_keyword.csv', encoding='utf-8-sig')

## 3. 데이터 전처리

actor_df.head(2)

webtoon_df.head(2)

### 1) actor data : Name, Age, Sex, activation_content, 배역내용, 제목, 내용, keywords만 가져온 데이터 프레임 생성

actor_df = actor_df[['Name', 'Age', 'Sex', 'activation_content', '배역내용', '제목, 내용', 'keywords']]

actor_df.head(2)

### 2) activation_content, 배역내용, 제목, 내용 합쳐서 새로운 columns 생성

# 전체 내용을 담을 list
actor_content_list = []

# 각 columns의 내용이 담긴 리스트 생성
activation_list = actor_df['activation_content'].to_list()
act_list = actor_df['배역내용'].to_list()
community_list = actor_df['제목, 내용'].to_list()

# 합치기
for i in range(len(activation_list)):
    actor_content_list.append(activation_list[i] + act_list[i] + community_list[i])

# 새로운 columns 생성
actor_df['content'] = actor_content_list

# 확인
actor_df.head(2)

# 필요없는 columns를 제거한 DataFrame 생성
actor_df = actor_df[['Name', 'Age', 'Sex', 'content', 'keywords']]

# 확인
actor_df.head(2)

## 4. 키워드 추출

### 1) 모델 로드

model = SentenceTransformer('sentence-transformers/xlm-r-100langs-bert-base-nli-stsb-mean-tokens')

### 2) 임베딩과 키워드 top5 추출

### (1) 배우 키워드 추출

# Nax Sum Similarity 함수 정의
def max_sum_sim(doc_embedding, candidate_embeddings, words, top_n, nr_candidates):
    # 문서와 각 키워드들 간의 유사도
    distances = cosine_similarity(doc_embedding, candidate_embeddings)

    # 각 키워드들 간의 유사도
    distances_candidates = cosine_similarity(candidate_embeddings, 
                                            candidate_embeddings)

    # 코사인 유사도에 기반하여 키워드들 중 상위 top_n개의 단어를 pick.
    words_idx = list(distances.argsort()[0][-nr_candidates:])
    words_vals = [words[index] for index in words_idx]
    distances_candidates = distances_candidates[np.ix_(words_idx, words_idx)]

    # 각 키워드들 중에서 가장 덜 유사한 키워드들간의 조합을 계산
    min_sim = np.inf
    candidate = None
    for combination in itertools.combinations(range(len(words_idx)), top_n):
        sim = sum([distances_candidates[i][j] for i in combination for j in combination if i != j])
        if sim < min_sim:
            candidate = combination
            min_sim = sim

    return [words_vals[idx] for idx in candidate]

# 배우 내용과 형용사 가져오기
actor_content_list = actor_df['content'].to_list()
actor_adjective_list = []

for i in actor_df['keywords']:
    actor_adjective_list.append(i.split(', '))

# 배우 키워드 추출
keywords = []

for i in tqdm_notebook(range(len(actor_content_list))):
    # 임베딩
    doc_embedding = model.encode([actor_content_list[i]])
    candidate_embeddings = model.encode(actor_adjective_list[i])
    
    # keyword 추출
    keyword = max_sum_sim(doc_embedding, candidate_embeddings, actor_adjective_list[i], 5, 10)
    keywords.append(", ".join(keyword))

# keyword columns 생성
actor_df['5_keywords'] = keywords
actor_df.head(2)

# Name, Age, Sex, 5_keywords columns 만 가지는 새로운 데이터 프레임 생성
actor_df = actor_df[['Name', 'Age', 'Sex', '5_keywords']]
actor_df.head(2)

# 데이터 프레임 저장
actor_df.to_csv('../text_data/actor_keywords.csv', encoding='utf-8-sig', index=False)

### (2) 웹툰 키워드 추출

# null 값 정리
webtoon_df = webtoon_df.dropna().reset_index(drop=True)

# 웹툰 등장인물 내용과 형용사 가져오기
webtoon_content_list = webtoon_df['설명'].to_list()
webtoon_adjective_list = []

for adjective in webtoon_df['keyword']:
    webtoon_adjective_list.append(adjective.split(', '))

# 웹툰 키워드 추출
keywords = []

for i in tqdm_notebook(range(len(webtoon_content_list))):
    # 임베딩
    doc_embedding = model.encode([webtoon_content_list[i]])
    candidate_embeddings = model.encode(webtoon_adjective_list[i])
    
    # keyword 추출
    try:
        keyword = max_sum_sim(doc_embedding, candidate_embeddings, webtoon_adjective_list[i], 5, 10)
        keywords.append(", ".join(keyword))
    except:
        keywords.append(np.nan)

# keyword columns 생성
webtoon_df['5_keywords'] = keywords
webtoon_df.head(2)

# Title, Character, Sex, Age, 5_keywords columns만 가지는 새로운 데이터 프레임 생성
webtoon_df = webtoon_df[['Title', 'Character', 'Sex', 'Age', '5_keywords']]
webtoon_df.head(2)

# 데이터 프레임 저장
webtoon_df.to_csv('../text_data/webtoon_keywords.csv', encoding='utf-8-sig', index=False)
