==> prototype_ver_01/5) Calculate image cosine similarity.ipynb <==
패키지 설치
!pip install img2vec_pytorch

## 1. 패키지 로드

import cv2
import numpy as np
import pandas as pd
from matplotlib import pyplot as plt
from tqdm import tqdm_notebook
import re
import dlib

from img2vec_pytorch import Img2Vec
from PIL import Image
from sklearn.metrics.pairwise import cosine_similarity

import warnings # 경고문 무시
warnings.filterwarnings('ignore')

## 2. 클래스 정의

class CImg:
    def __init__(self):
        '''init class'''
        self.img2vec = Img2Vec(cuda=True)
  
    def img_to_rgb(self, image):
        '''convert image to rgb'''
        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)

        return image

    def img_to_bgr(self, image):
        '''convert image to bgr'''
        image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)

        return image

    def img_adjust(self, image, brightness = 0, contrast = 30):
        '''adjust image'''
        image = np.int16(image)
        image = image * (contrast / 127 + 1) - contrast + brightness
        image = np.clip(image, 0, 255)
        image = np.uint8(image)

        return image

    def img_resize(self, image, size = 256):
        '''resize image'''
        image = cv2.resize(image, (size, size))

        return image
    
    def pil_to_cv(self, pil_image):
        '''PIL -> cv2'''
        numpy_image = np.array(pil_image)
        cv_image = cv2.cvtColor(numpy_image, cv2.COLOR_RGB2BGR)

        return cv_image
    
    def cv_to_pil(self, cv_image):
        '''cv2 -> PIL'''
        color_coverted = cv2.cvtColor(cv_image, cv2.COLOR_BGR2RGB)
        pil_image = Image.fromarray(color_coverted)
        return pil_image
    
    def img_face_align(self, image, size = 256, padding = 0.65):
        '''face align image'''
        detector = dlib.get_frontal_face_detector()
        sp = dlib.shape_predictor('../util/shape_predictor_5_face_landmarks.dat')
        dets = detector(image)
        if dets:
            pass
        else:
            #print("No detect face")
            return image
        s = sp(image, dets[0])
        image = dlib.get_face_chip(image, s, size=size, padding=padding)

        return image

    def img_compare(self, image_1, image_2):
        '''compare image'''
        image_1_vec = self.img2vec.get_vec(image_1, tensor=False)
        image_2_vec = self.img2vec.get_vec(image_2, tensor=False)
        
        image_1_vec = image_1_vec.reshape(1, -1)
        image_2_vec = image_2_vec.reshape(1, -1)
        
        return cosine_similarity(image_1_vec, image_2_vec)[0][0]


    def run_image_compare(self, image_1, image_2, size = 256, padding = 0.65, ratio = 0.75, brightness = 0, contrast = 30, show=False):
        '''compare image'''
        image_1 = self.img_to_rgb(image_1)
        image_1 = self.img_adjust(image_1, brightness=brightness, contrast=contrast)
        image_1 = self.img_resize(image_1, size=size)
        image_1 = self.img_face_align(image_1, size=size, padding=padding)
        image_1 = self.img_to_bgr(image_1)
        image_1 = self.cv_to_pil(image_1)
        
        image_2 = self.img_to_rgb(image_2)
        image_2 = self.img_adjust(image_2, brightness=brightness, contrast=contrast)
        image_2 = self.img_resize(image_2, size=size)
        image_2 = self.img_face_align(image_2, size=size, padding=padding)
        image_2 = self.img_to_bgr(image_2)
        image_2 = self.cv_to_pil(image_2)
        
        similarity = self.img_compare(image_1, image_2)

        return similarity

    def run(self, actor, webtoon, actor_image_path, webtoon_image_path):
        '''calculate all actor images and webtoon images similarity '''
        actor_df = actor.copy()
        webtoon_df = webtoon.copy()
        
        # webtoon name list and actor name list
        webtoon_names = webtoon_df['Title'].unique()
        actor_name_list = actor_df['Name'].to_list()
        actor_male_name_list = actor_df[actor_df['Sex'] == '남'].Name.to_list()
        actor_female_name_list = actor_df[actor_df['Sex'] == '여'].Name.to_list()
        
        # result dataframe
        result_df = actor_df[['Name', 'Age', 'Sex']]
        
        for webtoon_name in tqdm_notebook(webtoon_names):
            try:
                # get characters's names from webtoon name
                charater_names = webtoon_df[webtoon_df['Title'] == webtoon_name]['Character'].to_list()

                # calculate actors's images and characters's images
                for charater_name in charater_names:
                    try:
                        # character's image load 
                        character_image = cv2.imread(f'{webtoon_image_path}/{webtoon_name}_{charater_name}.jpg', flags=cv2.IMREAD_COLOR)

                        # get character's sex, age
                        character_sex = webtoon_df[webtoon_df['Character'] == charater_name]['Sex'].values[0]
                        character_age = webtoon_df[webtoon_df['Character'] == charater_name]['Age'].values[0]
                        character_age = int(re.sub('[^0-9]', '', character_age))
                    
                        # calculation according to sex
                        if character_sex == '남':
                            for actor_name in actor_male_name_list:
                                actor_age = int(actor_df[actor_df['Name'] == actor_name]['Age'].values[0])
                                if (character_age < actor_age) and (character_age + 20 > actor_age):
                                    actor_image = cv2.imread(f'{actor_image_path}/webtoon_{actor_name}.jpg', flags=cv2.IMREAD_COLOR)
                                    n_sim = self.run_image_compare(actor_image, character_image)
                                    result_df.loc[(result_df['Name'] == actor_name), f'{webtoon_name}_{charater_name}_image_score'] = n_sim
                        elif character_sex == '여':
                            for actor_name in actor_female_name_list:
                                actor_age = int(actor_df[actor_df['Name'] == actor_name]['Age'].values[0])
                                if (character_age < actor_age) and (character_age + 20 > actor_age):
                                    actor_image = cv2.imread(f'{actor_image_path}/webtoon_{actor_name}.jpg', flags=cv2.IMREAD_COLOR)
                                    n_sim = self.run_image_compare(actor_image, character_image)
                                    result_df.loc[(result_df['Name'] == actor_name), f'{webtoon_name}_{charater_name}_image_score'] = n_sim
                        else: # if, sex == NaN
                            for actor_name in actor_name_list:
                                actor_age = int(actor_df[actor_df['Name'] == actor_name]['Age'].values[0])
                                if (character_age < actor_age) and (character_age + 20 > actor_age):
                                    actor_image = cv2.imread(f'{actor_image_path}/webtoon_{actor_name}.jpg', flags=cv2.IMREAD_COLOR)
                                    n_sim = self.run_image_compare(actor_image, character_image)
                                    result_df.loc[(result_df['Name'] == actor_name), f'{webtoon_name}_{charater_name}_image_score'] = n_sim
                    except:
                        print(f'{charater_name} Error')
            except Exception as e:
                print(e)
                print(f'{webtoon_name} Error')
                
        
        return result_df

 ## 3. 이미지 로드 및 테스트

actor_image_path = '../image_data/Actor2Webtoon'
webtoon_image_path = '../image_data/Webtoon'

actor_image = cv2.imread(f'{actor_image_path}/webtoon_감우성.jpg')
webtoon_image = cv2.imread(f'{webtoon_image_path}/폭풍의전학생_주인공.jpg')

img = CImg()

img.run_image_compare(actor_image, webtoon_image)

## 4. for문을 통해 등장인물 - 배우 간 모든 이미지 유사도

actor_data = pd.read_csv('../text_data/Actor.csv', encoding='utf-8-sig')
webtoon_data = pd.read_csv('../text_data/Webtoon.csv', encoding='utf-8-sig')

result_df = img.run(actor_data, webtoon_data, actor_image_path, webtoon_image_path)

result_df.to_csv('../text_data/cos_sim_score_df.csv', encoding='utf-8-sig', index=False)

result_df.head()
